compiled yarp2 (no drivers yet) with visual studio 8, fixed complaints, requires compiling ace with a pragma
svn path=/trunk/yarp2/; revision=3390

+++ src/libYARP_sig_compat/src/YARPSVD.cpp
@@ -37,7 +37,7 @@
 ///
 
 ///
-/// $Id: YARPSVD.cpp,v 1.2 2006-05-15 15:57:59 eshuy Exp $
+/// $Id: YARPSVD.cpp,v 1.3 2006-05-21 13:48:22 gmetta Exp $
 ///
 ///
 
@@ -58,10 +58,10 @@
 // 
 //
 //
-static float at,bt,ct;
+static double at,bt,ct;
 #define PYTHAG(a,b) ((at=fabs(a)) > (bt=fabs(b)) ? (ct=bt/at,at*sqrt(1.0+ct*ct)) : (bt ? (ct=at/bt,bt*sqrt(1.0+ct*ct)): 0.0))
 
-static float maxarg1,maxarg2;
+static double maxarg1,maxarg2;
 #define MYMAX(a,b) (maxarg1=(a),maxarg2=(b),(maxarg1) > (maxarg2) ?	(maxarg1) : (maxarg2))
 
 #define SIGN(a,b) ((b) >= 0.0 ? fabs(a) : -fabs(a))
@@ -336,7 +336,7 @@ void SvdSolve(const YMatrix& u,
     int n = u.NCols ();
         
 	int jj,j,i;
-	float s;
+	double s;
 
     YVector tmp (n);
